Sandbox: new module (early-stage) and lots of fixes
There is a lot of work to do for this to be fully functional, but I decided to
commit it for those interested.

Also cleaned up a lot of stuff, most importantly:
	1. Host/Origin cr3 usage for Windows
	2. EPTP list usage

Now it's a lot more cleaner for allocating/using EPT pointers.

And now subverting is up to a user mode process instead of automatic on driver
load.

Signed-off-by: Ahmed Samy <asamy@protonmail.com>